Added indicator for boss monsters in combat- and slayer areas
