+++ src/Constants.ts
@@ -6,8 +6,8 @@ export class Constants {
 
     static readonly MISSING_ARTWORK_URL = "https://cdn.melvor.net/core/v018/assets/media/main/missing_artwork.png";
 
-    static readonly MONSTER_TYPE_INDICATOR_BADGE_CLASS = "customModifiersInMelvor__monster-type-indicator-badge";
-    static readonly MONSTER_TYPE_INDICATOR_BADGE_BR_CLASS = "customModifiersInMelvor__monster-type-indicator-badge-br";
+    static readonly COMBAT_AREAS_INDICATOR_BADGE_CLASS = "customModifiersInMelvor__monster-type-indicator-badge";
+    static readonly COMBAT_AREAS_INDICATOR_BADGE_BR_CLASS = "customModifiersInMelvor__monster-type-indicator-badge-br";
 
     // Based on the Typescript type "AttackType" which unfortunately isn't a usable enum
     static readonly ATTACK_TYPES_MELEE = "melee";
